[CLI] - Moving Commands Under Create
Moving a couple commands under create to reduce the verbage
